Update PERCPU map

+++ xdp_prog/common_kern_user.h
@@ -26,8 +26,8 @@ typedef __u64               fixed;
 typedef struct {
     __u64 time_in;
     __u64 proc_time;  /*proc_time += time_out - time_in*/
-    __u32 total_pkts;
-    __u32 total_bytes;
+    __u64 total_pkts;
+    __u64 total_bytes;
 } accounting;
 
 /* Flow key structure */
